handle extension version update

+++ ExtensionHandler/Windows/src/bin/uninstall.ps1
@@ -16,11 +16,13 @@ if (!(Test-Path variable:PSScriptRoot) -or !($PSScriptRoot)) { # $PSScriptRoot i
 Import-Module $PSScriptRoot\AzureExtensionHandler.psm1
 Import-Module $PSScriptRoot\RMExtensionHandler.psm1 -DisableNameChecking
 Import-Module $PSScriptRoot\RMExtensionStatus.psm1
+"$PSScriptRoot\Constants.ps1"
 
 Initialize-ExtensionLogFile
 $config = Get-ConfigurationFromSettings
 $configuredAgentExists = Test-AgentAlreadyExists $config
-if($configuredAgentExists)
+$isUpdateExtensionScenario = Test-Path "$agentWorkingFolder\$updateFileName"
+if($configuredAgentExists -and !$isUpdateExtensionScenario)
 {
     Remove-Agent $config
 }
